Refactor training and employee data handling

+++ main.py
@@ -1,9 +1,10 @@
 from fastapi import FastAPI
-from pydantic import BaseModel ,Field
+from pydantic import BaseModel ,Field,EmailStr
 from datetime import date
 import uuid
 import sqlite3
 import logging
+
 
 logging.basicConfig(
     filename="train.log",
@@ -57,15 +58,15 @@ class training_data(BaseModel):
     trainer_name: str = Field(...,min_length=3,max_length=30)
     topic : str = Field(...,min_length=1,max_length=100)
     duration : float = Field(...,gt=0.1 , lt= 9)
-    description : str = Field(...,min_length=1,max_length=1500)
+    description : str = Field("Description Not Added",min_length=2,max_length=1500)
     training_date: date
 
 class Employee(BaseModel):
     employee_id:str = Field(...,min_length=6,max_length=6)
     employee_name: str = Field(..., min_length=2,max_length=30)
     department: str = Field(...,min_length=1, max_length=20)
-    designation: str = Field(...,min_length=2, max_length=20)
-    email: str = Field(...,min_length=5, max_length=50)
+    designation: str = Field("Designation Not Mentioned",min_length=2, max_length=20)
+    email:EmailStr
 
 @app.post("/add_training")
 def create_training_sheet(payload:training_data):
@@ -85,14 +86,14 @@ def create_training_sheet(payload:training_data):
                    training_date) VALUES (?,?, ?, ?, ?, ?, ?, ?)
     """,
     (   
-        payload.employee_id,
-        unique_response_id,
-        payload.name,
-        payload.trainer_name,
-        payload.topic,
+        payload.employee_id.strip(),
+        unique_response_id.strip(),
+        payload.name.strip(),
+        payload.trainer_name.strip(),
+        payload.topic.strip(),
         round(payload.duration,1),
-        payload.description,
-        str(payload.training_date)
+        payload.description.strip(),
+        str(payload.training_date).strip()
     ))
     
     conn.commit()
@@ -132,19 +133,34 @@ def delete_trainingdata(response_id: str):
     conn.row_factory = sqlite3.Row 
 
     cursor = conn.cursor()
-
     cursor.execute(
-        "DELETE FROM training WHERE response_id=?",
+        "SELECT response_id FROM training WHERE response_id = ?",
         (response_id,)
     )
-    conn.commit()
-    conn.close()
-    logging.info(f"Training Data Deleted : {response_id}")
 
-    return {
-        "result": "Training Data Deleted Successfully"
-    }
+    existing = cursor.fetchone()
 
+    if existing:
+        cursor.execute(
+            "DELETE FROM training WHERE response_id=?",
+            (response_id,)
+        )
+        conn.commit()
+        conn.close()
+
+        logging.info(f"Training Data Deleted : {response_id}")
+
+        return {
+            "result": "Training Data Deleted Successfully"
+        }
+    else:
+        conn.close()
+
+        logging.info(f"Training Data Not Found : {response_id}")
+
+        return {
+            "result": "Training Data Not found"
+        }
 
     
 
@@ -178,11 +194,11 @@ def create_employee(payload: Employee):
         VALUES (?, ?, ?, ?, ?)
     """,
     (
-        payload.employee_id,
-        payload.employee_name,
-        payload.department,
-        payload.designation,
-        payload.email
+        payload.employee_id.strip(),
+        payload.employee_name.strip(),
+        payload.department.strip(),
+        payload.designation.strip(),
+        payload.email.strip()
     ))
 
     conn.commit()
@@ -205,33 +221,47 @@ def get_employee(employee_id: str):
         (employee_id,)
     )
 
-    record = cursor.fetchone()
+    record = dict(cursor.fetchone())
+
 
     conn.close()
 
     if not record:
         return {"message": "Record not found"}
     logging.warning(f"Data fetch for employee id {employee_id}")
+    def masking(data):
+        masking_data = data[0] + "*" * (len(data)-2) + data[len(data)-1]
+        return masking_data
+    record["email"] = masking(record["email"])
     return record
 
 @app.delete("/delete_emp/{employee_id}")
 def delete_emp(employee_id: str):
     conn = sqlite3.connect(DB_FILE)
     cursor = conn.cursor()
-    cursor.execute("""
-                    DELETE FROM employee WHERE employee_id=?
-                   """,
-                   (employee_id,)
-                    )
-    conn.commit()
-    conn.close()
-    logging.info(f"Employee Data Deleted : {employee_id}")
+    cursor.execute(
+        "SELECT employee_id FROM employee WHERE employee_id = ?",
+        (employee_id,)
+    )
 
-    return {
-        "result": "Employee Data Deleted Successfully"
-    }
+    existing = cursor.fetchone()
+    if existing:
+        cursor.execute("""
+                        DELETE FROM employee WHERE employee_id=?
+                    """,
+                    (employee_id,)
+                        )
+        conn.commit()
+        conn.close()
+        logging.info(f"Employee Data Deleted : {employee_id}")
 
+        return {
+            "result": "Employee Data Deleted Successfully"
+        }
+    else:
+        conn.close()
+        logging.info(f"Employee Data Not Fount : {employee_id}")
 
-
-
-
+        return {
+            "result": "Employee Data Not Found"
+        }
